Move task creation and configuration code to convention
This allows the tasks in question to be available and readable
immediately at the time of plugin application, outside
of afterEvaluate { }.

BREAKING CHANGES:
- Use plugin conventions for backwards compatibility
  with existing closure-based `distribution { }` syntax.
- Restricts configuration to one-time event.
- Distribution-related tasks are not available before calling
  `distribution { }`.
